python-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
joblib==1.5.2
lxml==6.0.1
multidict==6.6.4
numpy==2.0.2
pandas==2.3.2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
